pyproject.toml
[project]
name = "fintech-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2.0",
    "jinja2>=3.1.5",
    "langchain>=0.3.18",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.5",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
    "tabulate>=0.9.0",
]
